Pass data to room page.
Pass data to accommodation_room page.
